requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
contextlib2==21.6.0
cryptography==3.4.8
cycler==0.10.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kiwisolver==1.3.2
matplotlib==3.4.3
MedPy==0.4.0
ml-collections==0.1.0
multidict==5.1.0
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.2
plotly==5.3.1
protobuf==3.17.3
pycparser==2.20
pydicom==2.2.1
pynrrd==0.4.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
scipy==1.7.1
SimpleITK==2.1.1
six==1.16.0
tenacity==8.0.1
tensorboardX==2.4
torch==1.4.0
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.5.0
-f https://download.pytorch.org/whl/torch_stable.html
tqdm==4.62.2
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.2
wincertstore==0.2
wslink==1.0.7
yarl==1.6.3
zope.interface==5.4.0
